refactor: refactored view adding OutputView methods
